+++ gateway/src/apicast/policy/jwt_parser/jwt_parser.lua
@@ -1,4 +1,4 @@
--- OpenID Connect Authentication policy
+-- JWT Parser policy
 -- It will verify JWT signature against a list of public keys
 -- discovered through OIDC Discovery from the IDP.
 
@@ -8,7 +8,7 @@ local oidc_discovery = require('resty.oidc.discovery')
 local http_authorization = require('resty.http_authorization')
 local resty_url = require('resty.url')
 local policy = require('apicast.policy')
-local _M = policy.new('oidc_authentication', 'builtin')
+local _M = policy.new('jwt_parser', 'builtin')
 
 local tostring = tostring
 
@@ -23,7 +23,7 @@ local function valid_issuer_endpoint(endpoint)
 end
 
 local new = _M.new
---- Initialize a oidc_authentication
+--- Initialize jwt_parser policy
 -- @tparam[opt] table config Policy configuration.
 function _M.new(config)
   local self = new(config)
@@ -42,53 +42,51 @@ local function bearer_token()
   return http_authorization.new(ngx.var.http_authorization).token
 end
 
-function _M:rewrite(context)
-  local access_token = bearer_token()
-
-  if access_token or self.required then
-    local jwt, err = self.oidc:parse(access_token)
-
-    if jwt then
-      context[self] = jwt
-      context.jwt = jwt
-    else
-      ngx.log(ngx.WARN, 'failed to parse access token ', access_token, ' err: ', err)
-    end
-  end
-end
-
 local function exit_status(status)
   ngx.status = status
   -- TODO: implement content negotiation to generate proper content with an error
   return ngx.exit(status)
 end
 
-local function challenge_response()
+local function challenge_response(status)
   ngx.header.www_authenticate = 'Bearer'
 
-  return exit_status(ngx.HTTP_UNAUTHORIZED)
+  return exit_status(status)
 end
 
-function _M:access(context)
-  local jwt = context[self]
+local function check_compatible(context)
+  local service = context.service or {}
+  local authentication = service.authentication_method or service.backend_version
+  if authentication == "oidc" or authentication == "oauth" then
+    ngx.log(ngx.WARN, 'jwt_parser is incompatible with OIDC authentication mode')
+    return false
+  end
+  return true
+end
 
-  if not jwt or not jwt.token then
+function _M:rewrite(context)
+  if not check_compatible(context) then
+    return
+  end
+
+  local access_token = bearer_token()
+
+  if not access_token then
     if self.required then
-      return challenge_response()
-    else
-      return
+        return challenge_response(context.service.auth_failed_status)
     end
   end
 
-  local ok, err = self.oidc:verify(jwt)
+  if access_token then
+    local _, _, jwt_payload, err = self.oidc:transform_credentials({access_token=access_token})
 
-  if not ok then
-    ngx.log(ngx.INFO, 'JWT verification error: ', err, ' token: ', tostring(jwt))
+    if err then
+      ngx.log(ngx.WARN, 'failed to parse access token ', access_token, ' err: ', err)
+      return exit_status(context.service.auth_failed_status)
+    end
 
-    return exit_status(ngx.HTTP_FORBIDDEN)
+    context.jwt = jwt_payload
   end
-
-  return ok
 end
 
 return _M